--- simulation-ros/src/turtlebot2i/turtlebot2i_mrcnn/mrcnn/MANIFEST.in
@@ -1,3 +1,0 @@
-include README.md
-include LICENSE
-include requirements.txt
