--- simulation-ros/src/turtlebot2i/turtlebot2i_mrcnn/mrcnn/process_video/README.md
@@ -1,2 +1,0 @@
-# Mask-RCNN-series
-This series we'll use Mask RCNN
