--- simulation-ros/src/turtlebot2i/turtlebot2i_mrcnn/mrcnn/process_video/visualize_cv2.py
@@ -1,167 +1,0 @@
-import cv2
-import numpy as np
-import os
-import sys
-# import coco
-# import utils
-
-ROOT_DIR = os.path.abspath("../")
-print("--------> ", ROOT_DIR)
-
-from mrcnn import model as modellib
-from mrcnn.config import Config
-
-# sys.path.append(os.path.join(ROOT_DIR, "/samples/coco/"))  # To find local version
-# from samples.coco import coco
-
-# ROOT_DIR = os.getcwd()
-MODEL_DIR = os.path.join(ROOT_DIR, "processVideo/logs")
-COCO_MODEL_PATH = os.path.join(ROOT_DIR, "mask_rcnn_coco.h5")
-TEST_MODEL_PATH = os.path.join(ROOT_DIR, "logs/imagenet_vrepall_1003.h5")
-
-#if not os.path.exists(COCO_MODEL_PATH):
-#    utils.download_trained_weights(COCO_MODEL_PATH)
-
-class ShapesConfig(Config):
-    """Configuration for training on the toy shapes dataset.
-    Derives from the base Config class and overrides values specific
-    to the toy shapes dataset.
-    """
-    # Give the configuration a recognizable name
-    NAME = "vrepAll"
-
-    # Train on 1 GPU and 8 images per GPU. We can put multiple images on each
-    # GPU because the images are small. Batch size is 8 (GPUs * images/GPU).
-    GPU_COUNT = 1
-    IMAGES_PER_GPU = 1
-
-    # Number of classes (including background)
-    NUM_CLASSES = 1 + 8  # background + 3 shapes
-
-    # Use small images for faster training. Set the limits of the small side
-    # the large side, and that determines the image shape.
-    IMAGE_MIN_DIM = 480
-    IMAGE_MAX_DIM = 640
-
-    # Use smaller anchors because our image and objects are small
-    RPN_ANCHOR_SCALES = (8 * 4, 16 * 4, 32 * 4, 64 * 4, 128 * 4)  # anchor side in pixels
-
-    # Reduce training ROIs per image because the images are small and have
-    # few objects. Aim to allow ROI sampling to pick 33% positive ROIs.
-    TRAIN_ROIS_PER_IMAGE = 40
-
-    # Use a small epoch since the data is simple
-    STEPS_PER_EPOCH = 100
-
-    # use small validation steps since the epoch is small
-    VALIDATION_STEPS = 5
-
-# class InferenceConfig(coco.CocoConfig):
-class InferenceConfig(ShapesConfig):
-    GPU_COUNT = 1
-    IMAGES_PER_GPU = 1
-
-
-config = InferenceConfig()
-config.display()
-
-model = modellib.MaskRCNN(
-    mode="inference", model_dir=MODEL_DIR, config=config
-)
-print("Load model from", MODEL_DIR)
-model.load_weights(TEST_MODEL_PATH, by_name=True)
-print("Load weight from ", TEST_MODEL_PATH)
-class_names = ['BG', 'SlidingDoor', 'Wall', 'Shelf', 'Robot', 'Human', 'ConveyorBelt', 'Dockstation', 'Product']
-# class_names = [
-#     'BG', 'person', 'bicycle', 'car', 'motorcycle', 'airplane',
-#     'bus', 'train', 'truck', 'boat', 'traffic light',
-#     'fire hydrant', 'stop sign', 'parking meter', 'bench', 'bird',
-#     'cat', 'dog', 'horse', 'sheep', 'cow', 'elephant', 'bear',
-#     'zebra', 'giraffe', 'backpack', 'umbrella', 'handbag', 'tie',
-#     'suitcase', 'frisbee', 'skis', 'snowboard', 'sports ball',
-#     'kite', 'baseball bat', 'baseball glove', 'skateboard',
-#     'surfboard', 'tennis racket', 'bottle', 'wine glass', 'cup',
-#     'fork', 'knife', 'spoon', 'bowl', 'banana', 'apple',
-#     'sandwich', 'orange', 'broccoli', 'carrot', 'hot dog', 'pizza',
-#     'donut', 'cake', 'chair', 'couch', 'potted plant', 'bed',
-#     'dining table', 'toilet', 'tv', 'laptop', 'mouse', 'remote',
-#     'keyboard', 'cell phone', 'microwave', 'oven', 'toaster',
-#     'sink', 'refrigerator', 'book', 'clock', 'vase', 'scissors',
-#     'teddy bear', 'hair drier', 'toothbrush'
-# ]
-
-
-def random_colors(N):
-    np.random.seed(1)
-    colors = [tuple(255 * np.random.rand(3)) for _ in range(N)]
-    return colors
-
-colors = random_colors(len(class_names))
-class_dict = {
-    name: color for name, color in zip(class_names, colors)
-}
-
-def apply_mask(image, mask, color, alpha=0.5):
-    """apply mask to image"""
-    for n, c in enumerate(color):
-        image[:, :, n] = np.where(
-            mask == 1,
-            image[:, :, n] * (1 - alpha) + alpha * c,
-            image[:, :, n]
-        )
-    return image
-
-def display_instances(image, boxes, masks, ids, names, scores):
-    """
-        take the image and results and apply the mask, box, and Label
-    """
-    n_instances = boxes.shape[0]
-
-    if not n_instances:
-        print('NO INSTANCES TO DISPLAY')
-    else:
-        assert boxes.shape[0] == masks.shape[-1] == ids.shape[0]
-
-    for i in range(n_instances):
-        if not np.any(boxes[i]):
-            continue
-
-        y1, x1, y2, x2 = boxes[i]
-        label = names[ids[i]]
-        color = class_dict[label]
-        score = scores[i] if scores is not None else None
-        caption = '{} {:.2f}'.format(label, score) if score else label
-        mask = masks[:, :, i]
-
-        image = apply_mask(image, mask, color)
-        image = cv2.rectangle(image, (x1, y1), (x2, y2), color, 2)
-        image = cv2.putText(
-            image, caption, (x1, y1), cv2.FONT_HERSHEY_COMPLEX, 0.7, color, 2
-        )
-
-    return image
-
-if __name__ == '__main__':
-    """
-        test everything
-    """
-
-    capture = cv2.VideoCapture(0)
-
-    # these 2 lines can be removed if you dont have a 1080p camera.
-    capture.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
-    capture.set(cv2.CAP_PROP_FRAME_HEIGHT, 480)
-
-    while True:
-        ret, frame = capture.read()
-        results = model.detect([frame], verbose=1)
-        r = results[0]
-        frame = display_instances(
-            frame, r['rois'], r['masks'], r['class_ids'], class_names, r['scores']
-        )
-#        cv2.imshow('frame', frame)
-        if cv2.waitKey(1) & 0xFF == ord('q'):
-            break
-
-    capture.release()
-    cv2.destroyAllWindows()
