--- simulation-ros/src/turtlebot2i/turtlebot2i_mrcnn/mrcnn/setup.cfg
@@ -1,4 +1,0 @@
-[metadata]
-description-file = README.md
-license-file = LICENSE
-requirements-file = requirements.txt
