+++ simulation-ros/src/turtlebot2i/turtlebot2i_mrcnn/mrcnn/tools/bag2cv.py
@@ -3,7 +3,6 @@
 
 # Extract images from a bag file.
 
-#PKG = 'beginner_tutorials'
 import roslib;   #roslib.load_manifest(PKG)
 import rosbag
 import rospy
@@ -15,30 +14,6 @@ from cv_bridge import CvBridgeError
 # Reading bag filename from command line or roslaunch parameter.
 #import os
 #import sys
-
-'''
-enwhsaa@elx77440xgh:~/GetTrain$ rostopic list
-/keyop/teleop
-/rosout
-/rosout_agg
-/tf
-/turtlebot2i/camera/cloud
-/turtlebot2i/camera/depth/raw_image
-/turtlebot2i/camera/rgb/raw_image
-/turtlebot2i/camera_sr300/cloud
-/turtlebot2i/camera_sr300/depth/raw_image
-/turtlebot2i/camera_sr300/rgb/raw_image
-/turtlebot2i/commands/motor_power
-/turtlebot2i/commands/velocity
-/turtlebot2i/events/bumper
-/turtlebot2i/events/cliff
-/turtlebot2i/events/wheel_drop
-/turtlebot2i/lidar/scan
-/turtlebot2i/odom
-/turtlebot2i/sensors/core
-/turtlebot2i/sensors/dock_ir
-
-'''
 
 class ImageCreator():
 
@@ -58,8 +33,6 @@ class ImageCreator():
                         cv2.imwrite(image_name, cv_image)  #save image
 
 if __name__ == '__main__':
-
-    #rospy.init_node(PKG)
 
     try:
         image_creator = ImageCreator()
